computer graphics: transformation

+++ build.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-rm -rf public.copy && cp -r public public.copy
+if [ -f public ]; then
+    rm -rf public.copy && cp -r public public.copy
+fi
 
 zola build && \
     echo "blog.kiyoko.io" > public/CNAME
