+++ serve.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-rm -rf public.copy && cp -r public public.copy
+if [ -f public ]; then
+    rm -rf public.copy && cp -r public public.copy
+fi
 
 zola serve
